added helge's fix to WORD
git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/PCSC@158 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/PCSC/wintypes.h
@@ -36,7 +36,7 @@ extern "C"
 	typedef const void *LPCVOID;
 	typedef unsigned long DWORD;
 	typedef unsigned long *PDWORD;
-	typedef DWORD WORD;
+        typedef unsigned short WORD;       /* H. Bragstad Schlumberger */
 	typedef long LONG;
 	typedef long RESPONSECODE;
 	typedef const char *LPCSTR;
